server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
av==10.0.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.2
ctranslate2==3.24.0
fastapi==0.109.2
faster-whisper==0.10.0
ffmpeg-python==0.2.0
filelock==3.13.1
firebase-admin==6.4.0
flatbuffers==23.5.26
fsspec==2024.2.0
future==0.18.3
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
mpmath==1.3.0
msgpack==1.0.7
numpy==1.26.3
onnxruntime==1.17.0
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tokenizers==0.15.1
tqdm==4.66.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
